make restart mechanism work with both pyinstaller and regular environments
