+++ pyproject.toml
@@ -83,5 +83,4 @@ update_images = "photomap.backend.imagetool:main"
 search_images = "photomap.backend.imagetool:main"
 search_text = "photomap.backend.imagetool:main"
 find_duplicate_images = "photomap.backend.imagetool:main"
-photomap_server = "photomap.backend.photomap_server:main"
-start_photomap = "photomap.backend.main_wrapper:start_photomap_loop"
+start_photomap = "photomap.backend.photomap_server:main"
